[SecurityBundle] added compiler pass for checking voters validity
