+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/Compiler/AddSecurityVotersPass.php
@@ -40,20 +40,6 @@ class AddSecurityVotersPass implements CompilerPassInterface
         if (!$voters) {
             throw new LogicException('No security voters found. You need to tag at least one with "security.voter"');
         }
-
-        foreach ($voters as $voter) {
-            $class = $container->getDefinition((string) $voter)->getClass();
-
-            if (!is_a($class, VoterInterface::class, true)) {
-                @trigger_error(sprintf('Using a security.voter tag on a class without implementing the %1$s is deprecated as of 3.4 and will be removed in 4.0. Implement the %1$s instead.', VoterInterface::class), E_USER_DEPRECATED);
-            }
-
-            if (!method_exists($class, 'vote')) {
-                // in case the vote method is completely missing, to prevent exceptions when voting
-                throw new LogicException(sprintf('%s should implement the %s interface when used as voter.', $class, VoterInterface::class));
-            }
-        }
-
         $adm = $container->getDefinition('security.access.decision_manager');
         $adm->replaceArgument(0, new IteratorArgument($voters));
     }
